Refactor - 1)Classes renamed.(2)Pacts organised based on caseworker and citizen
