+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/DivorceCaseMaintenanceReadForCaseWorker.java
@@ -1,0 +1,104 @@
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
+
+import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
+import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+
+import java.io.IOException;
+import java.util.Map;
+
+import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
+import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
+import au.com.dius.pact.consumer.junit5.PactTestFor;
+import au.com.dius.pact.core.model.RequestResponsePact;
+import au.com.dius.pact.core.model.annotations.Pact;
+import org.apache.http.client.fluent.Executor;
+import org.json.JSONException;
+import org.junit.After;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+@ExtendWith(PactConsumerTestExt.class)
+@ExtendWith(SpringExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@PactTestFor(providerName = "ccd", port = "8891")
+@SpringBootTest({
+    "core_case_data.api.url : localhost:8891"
+})
+public class DivorceCaseMaintenanceReadForCaseWorker {
+
+    public static final String SOME_AUTHORIZATION_TOKEN = "Bearer UserAuthToken";
+    public static final String SOME_SERVICE_AUTHORIZATION_TOKEN = "ServiceToken";
+
+    @Autowired
+    private CoreCaseDataApi coreCaseDataApi;
+
+    @Value("${ccd.jurisdictionid}")
+    String jurisdictionId;
+
+    @Value("${ccd.casetype}")
+    String caseType;
+
+    private static final String USER_ID ="123456";
+    private static final Long CASE_ID = 2000l;
+    private static final String SERVICE_AUTHORIZATION = "ServiceAuthorization";
+
+    @BeforeEach
+    public void setUpEachTest() throws InterruptedException {
+        Thread.sleep(2000);
+    }
+
+    @After
+    void teardown() {
+        Executor.closeIdleConnections();
+    }
+
+    @Pact(provider = "ccd", consumer = "divorce_caseMaintenanceService_caseworker")
+    RequestResponsePact readForCaseDetails(PactDslWithProvider builder) {
+        // @formatter:off
+        return builder
+            .given("Read For Caseworker")
+            .uponReceiving("A Read For CaseWorker is received.")
+            .path("/caseworkers/"
+                + USER_ID +
+                "/jurisdictions/" + jurisdictionId
+                + "/case-types/"
+                + caseType
+                + "/cases/"
+                +  CASE_ID)
+            .method("GET")
+            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
+                SOME_SERVICE_AUTHORIZATION_TOKEN)
+            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+            .willRespondWith()
+            .matchHeader(HttpHeaders.CONTENT_TYPE, "\\w+\\/[-+.\\w]+;charset=(utf|UTF)-8")
+            .status(200)
+            .body(buildCaseDetailsDsl(CASE_ID, "emailAddress@email.com",false, false))
+            .toPact();
+    }
+
+    @Test
+    @PactTestFor(pactMethod = "readForCaseDetails")
+    public void verifyReadForCaseDetails() throws IOException, JSONException {
+
+        CaseDetails caseDetailsReponse = coreCaseDataApi.readForCaseWorker(SOME_AUTHORIZATION_TOKEN,
+            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
+            caseType,CASE_ID.toString());
+        Map<String,Object> dataMap = caseDetailsReponse.getData() ;
+        assertThat(dataMap.get("outsideUKGrantCopies"), is(6));
+        assertThat(dataMap.get("primaryApplicantForenames"), is("Jon"));
+    }
+
+}
