+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/DivorceCaseMaintenanceStartEventForCaseWorker.java
@@ -1,27 +1,19 @@
 package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
 
-import static io.pactfoundation.consumer.dsl.LambdaDsl.newJsonBody;
-import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailList;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildStartEventReponse;
 
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 
 import java.io.IOException;
-import java.util.Map;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.hamcrest.Matcher;
 import org.json.JSONException;
 import org.junit.Before;
 import org.junit.jupiter.api.BeforeEach;
@@ -35,10 +27,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-//import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.domain.model.CcdCaseProperties.D8_PETITIONER_EMAIL;
-//import au.com.dius.pact.core.model.RequestResponsePact;
-//import au.com.dius.pact.core.model.annotations.Pact;
-
 @ExtendWith(PactConsumerTestExt.class)
 @ExtendWith(SpringExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -46,17 +34,13 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @SpringBootTest({
     "core_case_data.api.url : localhost:8891"
 })
-public class DivorceCaseMaintenance_StartForCaseWorker {
+public class DivorceCaseMaintenanceStartEventForCaseWorker {
 
     public static final String SOME_AUTHORIZATION_TOKEN = "Bearer UserAuthToken";
     public static final String SOME_SERVICE_AUTHORIZATION_TOKEN = "ServiceToken";
-    public static final String REGEX_DATE = "^((19|2[0-9])[0-9]{2})-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])$";
 
     @Autowired
     private CoreCaseDataApi coreCaseDataApi;
-
-    @Autowired
-    ObjectMapper objectMapper;
 
     @Value("${ccd.jurisdictionid}")
     String jurisdictionId;
@@ -68,12 +52,8 @@ public class DivorceCaseMaintenance_StartForCaseWorker {
     String createEventId;
 
     private static final String USER_ID ="123456";
-    private static final String CASE_ID = "654321";
+    private static final String CASE_ID = "2000";
     private static final String SERVICE_AUTHORIZATION = "ServiceAuthorization";
-    private CaseDataContent caseDataContent;
-    private CaseDetails caseDetails;
-    private static final String  ALPHABETIC_REGEX = "[/^[A-Za-z]+$/]+";
-
 
     @Before
     public void setUp() throws Exception {
@@ -85,51 +65,43 @@ public class DivorceCaseMaintenance_StartForCaseWorker {
         Thread.sleep(2000);
     }
 
-
-    @Pact(provider = "ccd", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact startForCaseWorker(PactDslWithProvider builder) {
+    @Pact(provider = "ccd", consumer = "divorce_caseMaintenanceService_caseworker")
+    RequestResponsePact startEventForCaseWorker(PactDslWithProvider builder) {
         // @formatter:off
         return builder
-            .given("A start event for caseworkder is  requested")
-            .uponReceiving("a request for a valid start event for caseworker")
+            .given("A StartEvent for Caseworkder is  requested")
+            .uponReceiving("A StartEvent for a caseworker is received.")
             .path("/caseworkers/" + USER_ID + "/jurisdictions/"
                 + jurisdictionId + "/case-types/"
                 + caseType
+                + "/cases/"
+                +  CASE_ID
                 + "/event-triggers/"
                 + createEventId
                 + "/token")
             .method("GET")
-            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
-                SOME_SERVICE_AUTHORIZATION_TOKEN)
+            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION, SOME_SERVICE_AUTHORIZATION_TOKEN)
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
             .matchHeader(HttpHeaders.CONTENT_TYPE, "\\w+\\/[-+.\\w]+;charset=(utf|UTF)-8")
             .status(200)
-            .body(buildStartEventReponse(createEventId , "token","someemailaddress.com", false,false))
+            .body(buildStartEventReponse("100", "testServiceToken" , "emailAddress@email.com", true, true))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "startForCaseWorker")
-    public void verifyStartEventForCitizen() throws IOException, JSONException {
+    @PactTestFor(pactMethod = "startEventForCaseWorker")
+    public void verifyStartEventForCaseworker() throws IOException, JSONException {
 
-        StartEventResponse startEventResponse = coreCaseDataApi.startForCaseworker(SOME_AUTHORIZATION_TOKEN,
-            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType,createEventId);
+        final StartEventResponse startEventResponse = coreCaseDataApi.startEventForCaseWorker(SOME_AUTHORIZATION_TOKEN,
+            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,caseType,CASE_ID,createEventId);
 
-        assertThat(startEventResponse.getEventId(), equalTo(createEventId));
-        assertThat(startEventResponse.getToken(), is("token"));
+        assertThat(startEventResponse.getEventId(), is("100"));
+        assertThat(startEventResponse.getToken(), is("testServiceToken"));
 
-        final Map<String,Object> caseData = startEventResponse.getCaseDetails().getData();
-
-        assertThat(caseData.get("outsideUKGrantCopies"), is(6));
-        // applicationType
-        assertThat(caseData.get("applicationType"), is("Personal"));
-
-
-//        assertThat(startEventResponse.getCaseDetails().getJurisdiction(), is("DIVORCE"));
-//        assertThat(startEventResponse.getCaseDetails().getCallbackResponseStatus(), is("DONE"));
-//        assertThat(startEventResponse.getCaseDetails().getCaseTypeId(), is("DIVORCE"));
-
+        assertThat(startEventResponse.getCaseDetails().getId(), is((2000L)));
+        assertThat(startEventResponse.getCaseDetails().getJurisdiction(), is("DIVORCE"));
+        assertThat(startEventResponse.getCaseDetails().getCallbackResponseStatus(), is("DONE"));
+        assertThat(startEventResponse.getCaseDetails().getCaseTypeId(), is("GRANT_OF_REPRESENTATION"));
     }
 }
