+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/DivorceCaseMaintenanceSubmitEventForCaseWorker.java
@@ -1,25 +1,22 @@
 package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
 
-import static java.lang.String.valueOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.ObjectMapperTestUtil.convertObjectToJsonString;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContent;
 
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.ResourceLoader;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.Before;
+import org.apache.http.client.fluent.Executor;
+import org.junit.After;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
@@ -38,16 +35,13 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @SpringBootTest({
     "core_case_data.api.url : localhost:8891"
 })
-public class DivorceCaseMaintenance_SubmitEventForCaseWorker {
+public class DivorceCaseMaintenanceSubmitEventForCaseWorker {
 
     public static final String SOME_AUTHORIZATION_TOKEN = "Bearer UserAuthToken";
     public static final String SOME_SERVICE_AUTHORIZATION_TOKEN = "ServiceToken";
 
     @Autowired
     private CoreCaseDataApi coreCaseDataApi;
-
-    @Autowired
-    ObjectMapper objectMapper;
 
     @Value("${ccd.jurisdictionid}")
     String jurisdictionId;
@@ -62,27 +56,24 @@ public class DivorceCaseMaintenance_SubmitEventForCaseWorker {
     private static final String CASE_ID = "2000";
     private static final String SERVICE_AUTHORIZATION = "ServiceAuthorization";
     private CaseDataContent caseDataContent;
-    private CaseDetails caseDetails;
-    private static final String ALPHABETIC_REGEX = "[/^[A-Za-z]+$/]+";
-    private static final String VALID_PAYLOAD_PATH = "json/base-case.json";
-
-
-    @Before
-    public void setUp() throws Exception {
-
-    }
 
     @BeforeEach
     public void setUpEachTest() throws InterruptedException {
         Thread.sleep(2000);
     }
 
-    @Pact(provider = "ccd", consumer = "divorce_caseMaintenanceService")
+    @After
+    void teardown() {
+        Executor.closeIdleConnections();
+    }
+
+
+    @Pact(provider = "ccd", consumer = "divorce_caseMaintenanceService_caseworker")
     RequestResponsePact submitEventForCaseWorker(PactDslWithProvider builder) throws Exception {
         // @formatter:off
         return builder
-            .given("A Submit event for Caseworkder  starts ")
-            .uponReceiving("When a Submit Event For Caseworker is triggered.")
+            .given("A SubmitEvent for Caseworkder is triggered")
+            .uponReceiving("A SubmitEvent For Case worker is received.")
             .path("/caseworkers/"+ USER_ID
                 + "/jurisdictions/" + jurisdictionId
                 + "/case-types/"    + caseType
@@ -96,7 +87,7 @@ public class DivorceCaseMaintenance_SubmitEventForCaseWorker {
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
             .status(200)
-            .body(buildCaseDetailsDsl(100L, "someemailaddress.com", false, false))
+            .body(buildCaseDetailsDsl(2000L, "someemailaddress.com", false, false))
             .matchHeader(HttpHeaders.CONTENT_TYPE, "\\w+\\/[-+.\\w]+;charset=(utf|UTF)-8")
             .toPact();
     }
@@ -110,7 +101,7 @@ public class DivorceCaseMaintenance_SubmitEventForCaseWorker {
         final CaseDetails caseDetails = coreCaseDataApi.submitEventForCaseWorker(SOME_AUTHORIZATION_TOKEN,
             SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId, caseType, CASE_ID,true,caseDataContent);
 
-        assertThat(caseDetails.getId(), is(100L));
+        assertThat(caseDetails.getId(), is(2000L));
         assertThat(caseDetails.getJurisdiction(), is("DIVORCE"));
         assertThat(caseDetails.getState(), is("CaseCreated"));
 
@@ -120,26 +111,4 @@ public class DivorceCaseMaintenance_SubmitEventForCaseWorker {
 
     }
 
-    private CaseDataContent getCaseDataContent() throws Exception {
-
-        final String caseData = ResourceLoader.loadJson(VALID_PAYLOAD_PATH);
-
-        final StartEventResponse startEventResponse = StartEventResponse.builder()
-            .eventId(createEventId)
-            .token(SOME_AUTHORIZATION_TOKEN)
-            .build();
-
-        final CaseDataContent caseDataContent = CaseDataContent.builder()
-            .eventToken(startEventResponse.getToken())
-            .event(
-                Event.builder()
-                    .id(startEventResponse.getEventId())
-                    .summary("divSummary")
-                    .description("div")
-                    .build()
-            ).data(convertObjectToJsonString(caseData))
-            .build();
-
-        return caseDataContent;
-    }
 }
