+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/util/ObjectMapperTestUtil.java
@@ -1,5 +1,9 @@
 package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util;
 
+import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
+import uk.gov.hmcts.reform.ccd.client.model.Event;
+import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 public class ObjectMapperTestUtil {
@@ -21,4 +25,6 @@ public class ObjectMapperTestUtil {
             throw new RuntimeException(e);
         }
     }
+
+
 }
